# backend/internal/extension/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentSkillInstallRequest": {
        "properties": {
          "characterId": {
            "type": "string"
          },
          "enable": {
            "default": false,
            "type": "boolean"
          },
          "previewId": {
            "type": "string"
          },
          "scope": {
            "enum": [
              "global",
              "character"
            ]
          }
        },
        "required": [
          "previewId",
          "scope"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "traceId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Amitia Extension Runtime API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/extensions/agent-skills": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "enum": [
                "global",
                "character"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "compatible",
                "compatible_with_warnings",
                "partially_compatible",
                "blocked"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent Skill page"
          }
        },
        "summary": "List visible Agent Skills"
      }
    },
    "/extensions/agent-skills/import/install": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSkillInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Installed, disabled by default"
          }
        },
        "summary": "Install a previously validated Agent Skill preview"
      }
    },
    "/extensions/agent-skills/import/preview": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "files": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "paths": {
                    "type": "string"
                  },
                  "rootName": {
                    "type": "string"
                  },
                  "source": {
                    "enum": [
                      "zip",
                      "directory"
                    ]
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server-validated import preview"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid or unsafe import"
          }
        },
        "summary": "Validate and preview a ZIP or directory Agent Skill import"
      }
    },
...
    "/extensions/agent-skills/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          }
        },
        "summary": "Disable and remove an Agent Skill"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent Skill detail"
          }
        },
        "summary": "Get Agent Skill details, compatibility and activation trace"
      }
    },
    "/extensions/agent-skills/{id}/activations": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activation history"
          }
        },
        "summary": "List redacted Agent Skill activation traces"
      }
    },
    "/extensions/agent-skills/{id}/assets/content": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sandboxed asset content"
          },
          "403": {
            "description": "Unsafe or out-of-scope asset"
          }
        },
        "summary": "Download a controlled non-executable Agent Skill asset"
      }
    },
    "/extensions/agent-skills/{id}/compatibility": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compatibility report"
          }
        },
        "summary": "Get Agent Skill compatibility report"
      }
    },
    "/extensions/agent-skills/{id}/disable": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disabled"
          }
        },
        "summary": "Disable an Agent Skill and evict prompt cache"
      }
    },
    "/extensions/agent-skills/{id}/enable": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enabled"
          }
        },
        "summary": "Enable a non-blocked Agent Skill"
      }
    },
    "/extensions/agent-skills/{id}/resources": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource index"
          }
        },
        "summary": "List Agent Skill resource metadata"
      }
    },
    "/extensions/agent-skills/{id}/resources/content": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource content"
          }
        },
        "summary": "Read or download a controlled Agent Skill text resource"
      }
    },
...
    "/extensions/canary/routes/{extensionId}": {
      "get": {
        "description": "Get canary route",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get canary route",
        "tags": [
          "Canary"
        ]
      }
    },
...
    "/extensions/capabilities": {
      "get": {
        "responses": {
          "200": {
            "description": "Capability definitions"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "List capability definitions"
      }
    },
...
    "/extensions/package-operations/{operationId}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "operationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation detail"
          },
          "404": {
            "description": "Operation not found"
          }
        },
        "summary": "Get a package operation audit record"
      }
    },
...
    "/extensions/packages/import/sessions/{sessionId}": {
      "get": {
        "parameters": [{"in": "path", "name": "sessionId", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Package preview session"}},
        "summary": "Get a package preview session"
      },
      "delete": {
        "parameters": [{"in": "path", "name": "sessionId", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Package preview session cancelled"}},
        "summary": "Cancel a package preview session"
      }
    },
    "/extension-package-operations/{operationId}": {
      "get": {
        "parameters": [{"in": "path", "name": "operationId", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Package operation"}},
        "summary": "Get a package operation compatibility view"
      }
    },
...
    "/extensions/packages/import/install": {
      "post": {
        "deprecated": true,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "410": {
            "description": "PACKAGE_INSTALL_ENDPOINT_RETIRED"
          }
        },
        "summary": "Retired package install endpoint",
        "x-replacement-endpoint": "/api/extensions/packages/artifacts"
      }
    },
    "/extensions/packages/import/preview": {
      "post": {
        "deprecated": true,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            },
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "scopeId": {
                    "type": "string"
                  },
                  "scopeType": {
                    "enum": [
                      "global",
                      "character"
                    ]
                  }
                },
                "required": [
                  "file",
                  "scopeType"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "410": {
            "description": "PACKAGE_PREVIEW_ENDPOINT_RETIRED"
          }
        },
        "summary": "Retired package preview endpoint",
        "x-replacement-endpoint": "/api/extensions/packages/artifacts"
      }
    },
...
    "/extensions/plugins": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin page"
          }
        },
        "summary": "List registered plugins"
      }
    },
    "/extensions/plugins/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Get plugin details"
      }
    },
    "/extensions/plugins/{id}/circuit/reset": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset"
          }
        },
        "summary": "Reset all plugin hook circuits"
      }
    },
    "/extensions/plugins/{id}/config": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Redacted config"
          }
        },
        "summary": "Get redacted plugin config"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid config"
          }
        },
        "summary": "Validate, encrypt and update plugin config"
      }
    },
    "/extensions/plugins/{id}/config/reset": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset"
          }
        },
        "summary": "Reset plugin config"
      }
    },
    "/extensions/plugins/{id}/disable": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disabled"
          }
        },
        "summary": "Disable plugin and its registered skills"
      }
    },
    "/extensions/plugins/{id}/enable": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enabled"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Incompatible or lifecycle failure"
          }
        },
        "summary": "Enable plugin and its registered skills"
      }
    },
    "/extensions/plugins/{id}/events": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "pending",
                "failed",
                "completed",
                "dead_letter"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "maximum": 100,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event delivery page"
          }
        },
        "summary": "List role-isolated plugin event deliveries"
      }
    },
    "/extensions/plugins/{id}/events/dead-letter": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dead-letter page"
          }
        },
        "summary": "List plugin dead letters"
      }
    },
    "/extensions/plugins/{id}/events/{eventId}/retry": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Queued"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Character scope mismatch"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not retryable"
          }
        },
        "summary": "Retry owned dead-letter event"
      }
    },
...
    "/extensions/plugins/{id}/permissions": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission grants"
          }
        },
        "summary": "Get role-isolated plugin capability grants"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "characterId": {
                    "type": "string"
                  },
                  "grants": {
                    "items": {
                      "$ref": "#/components/schemas/PermissionGrant"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "grants"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Denied"
          }
        },
        "summary": "Replace plugin capability grants"
      }
    },
    "/extensions/plugins/{id}/reload": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reloaded"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Reload failed"
          }
        },
        "summary": "Reload plugin with rollback on failure"
      }
    },
    "/extensions/plugins/{id}/schedules": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedules"
          }
        },
        "summary": "List role-isolated plugin schedules"
      }
    },
...
    "/extensions/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Uninstall result"
          },
          "409": {
            "description": "Dependency in use"
          }
        },
        "summary": "Uninstall a local Skill while preserving audits and runs"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/api"
    }
  ]
}